+++ lib/src/analytics/analytics_service.dart
@@ -24,7 +24,8 @@ class AnalyticsService {
   /// [_crashReportsInterface].
   void userId({required String userId}) {
     _eventBus.tryAddAnalytic(Loglytics._analyticsInterface?.setUserId(userId));
-    _eventBus.tryAddCrashReport(Loglytics._crashReportsInterface?.setUserIdentifier(userId));
+    _eventBus.tryAddCrashReport(
+        Loglytics._crashReportsInterface?.setUserIdentifier(userId));
     _loglytics?.log.analytic(name: 'user_id', value: userId);
   }
 
@@ -35,9 +36,10 @@ class AnalyticsService {
   void userProperty({required String property, required Object? value}) {
     final _value = value?.toString() ?? '';
     if (_value.isNotEmpty) {
-      _eventBus.tryAddAnalytic(
-          Loglytics._analyticsInterface?.setUserProperty(name: property, value: _value));
-      _eventBus.tryAddCrashReport(Loglytics._crashReportsInterface?.setCustomKey(property, _value));
+      _eventBus.tryAddAnalytic(Loglytics._analyticsInterface
+          ?.setUserProperty(name: property, value: _value));
+      _eventBus.tryAddCrashReport(
+          Loglytics._crashReportsInterface?.setCustomKey(property, _value));
       _loglytics?.log.analytic(name: '[PROPERTY] $property', value: _value);
     } else {
       _loglytics?.log.error('Refused setting empty value for $property');
@@ -45,7 +47,8 @@ class AnalyticsService {
   }
 
   /// Main method used for sending for the more flexible [CustomAnalytic]s.
-  void custom({required CustomAnalytic analytic}) => _logCustomAnalytic(analytic);
+  void custom({required CustomAnalytic analytic}) =>
+      _logCustomAnalytic(analytic);
 
   /// Sends an [AnalyticsTypes.tapped] based on given [subject] and possible [parameters].
   void tapped({
@@ -320,7 +323,9 @@ class AnalyticsService {
       parameters: parameters,
       type: AnalyticsTypes.input,
     );
-    if (_firstInput == null || !onlyFirstValue || !analytic.equals(_firstInput)) {
+    if (_firstInput == null ||
+        !onlyFirstValue ||
+        !analytic.equals(_firstInput)) {
       _logAnalytic(analytic);
     }
     _firstInput = analytic;
@@ -1028,17 +1033,32 @@ class AnalyticsService {
         ),
       );
 
+  /// Sends an [AnalyticsTypes.given] based on given [subject] and possible [parameters].
+  void given({
+    required String subject,
+    Map<String, Object?>? parameters,
+  }) =>
+      _logAnalytic(
+        Analytic(
+          subject: subject,
+          parameters: parameters,
+          type: AnalyticsTypes.given,
+        ),
+      );
+
   /// Sends the current screen based on given [subject] and possible [parameters].
   void screen({
     required String subject,
   }) {
     final name = subject;
-    _eventBus.tryAddAnalytic(Loglytics._analyticsInterface?.setCurrentScreen(name: name));
+    _eventBus.tryAddAnalytic(
+        Loglytics._analyticsInterface?.setCurrentScreen(name: name));
     _loglytics?.log.analytic(name: '[SCREEN] $name');
   }
 
   /// Resets all current analytics data.
-  Future<void> resetAnalytics() async => Loglytics._analyticsInterface?.resetAnalyticsData();
+  Future<void> resetAnalytics() async =>
+      Loglytics._analyticsInterface?.resetAnalyticsData();
 
   /// Resets the [_firstInput] used by [AnalyticsService.input].
   void resetFirstInput() => _firstInput = null;
@@ -1047,8 +1067,8 @@ class AnalyticsService {
   void _logAnalytic(Analytic analytic) {
     final name = analytic.name;
     final parameters = analytic.parameters;
-    _eventBus.tryAddAnalytic(
-        Loglytics._analyticsInterface?.logEvent(name: name, parameters: parameters));
+    _eventBus.tryAddAnalytic(Loglytics._analyticsInterface
+        ?.logEvent(name: name, parameters: parameters));
     _loglytics?.log.analytic(name: name, parameters: parameters);
   }