+++ lib/src/enums/analytics_types.dart
@@ -74,169 +74,97 @@ enum AnalyticsTypes {
   shown,
   loaded,
   saved,
-  notFound,
   found,
   completed,
   error,
+  given,
+  notFound,
   none,
 }
 
 /// Used to generate the proper String format when sending analytics to the analytics provider.
 extension AnalyticsTypesHelpers on AnalyticsTypes {
-  String get name {
+  String get value {
     switch (this) {
       case AnalyticsTypes.event:
-        return 'event';
       case AnalyticsTypes.tapped:
-        return 'tapped';
       case AnalyticsTypes.clicked:
-        return 'clicked';
       case AnalyticsTypes.focussed:
-        return 'focussed';
       case AnalyticsTypes.selected:
-        return 'selected';
       case AnalyticsTypes.connected:
-        return 'connected';
       case AnalyticsTypes.disconnected:
-        return 'disconnected';
       case AnalyticsTypes.viewed:
-        return 'viewed';
       case AnalyticsTypes.hidden:
-        return 'hidden';
       case AnalyticsTypes.opened:
-        return 'opened';
       case AnalyticsTypes.closed:
-        return 'closed';
       case AnalyticsTypes.failed:
-        return 'failed';
       case AnalyticsTypes.succeeded:
-        return 'succeeded';
       case AnalyticsTypes.sent:
-        return 'sent';
       case AnalyticsTypes.received:
-        return 'received';
       case AnalyticsTypes.validated:
-        return 'validated';
       case AnalyticsTypes.invalidated:
-        return 'invalidated';
       case AnalyticsTypes.searched:
-        return 'searched';
       case AnalyticsTypes.liked:
-        return 'liked';
       case AnalyticsTypes.shared:
-        return 'shared';
       case AnalyticsTypes.commented:
-        return 'commented';
       case AnalyticsTypes.input:
-        return 'input';
       case AnalyticsTypes.incremented:
-        return 'incremented';
       case AnalyticsTypes.decremented:
-        return 'decremented';
       case AnalyticsTypes.accepted:
-        return 'accepted';
       case AnalyticsTypes.declined:
-        return 'declined';
       case AnalyticsTypes.alert:
-        return 'alert';
       case AnalyticsTypes.scrolled:
-        return 'scrolled';
       case AnalyticsTypes.started:
-        return 'started';
       case AnalyticsTypes.stopped:
-        return 'stopped';
       case AnalyticsTypes.initialised:
-        return 'initialised';
       case AnalyticsTypes.disposed:
-        return 'disposed';
       case AnalyticsTypes.fetched:
-        return 'fetched';
       case AnalyticsTypes.set:
-        return 'set';
       case AnalyticsTypes.get:
-        return 'get';
       case AnalyticsTypes.foreground:
-        return 'foreground';
       case AnalyticsTypes.background:
-        return 'background';
       case AnalyticsTypes.purchased:
-        return 'purchased';
       case AnalyticsTypes.dismissed:
-        return 'dismissed';
       case AnalyticsTypes.upgraded:
-        return 'upgraded';
       case AnalyticsTypes.downgraded:
-        return 'downgraded';
       case AnalyticsTypes.interaction:
-        return 'interaction';
       case AnalyticsTypes.query:
-        return 'query';
       case AnalyticsTypes.confirmed:
-        return 'confirmed';
       case AnalyticsTypes.canceled:
-        return 'canceled';
       case AnalyticsTypes.created:
-        return 'created';
       case AnalyticsTypes.read:
-        return 'read';
       case AnalyticsTypes.updated:
-        return 'updated';
       case AnalyticsTypes.deleted:
-        return 'deleted';
       case AnalyticsTypes.added:
-        return 'added';
       case AnalyticsTypes.removed:
-        return 'removed';
       case AnalyticsTypes.subscribed:
-        return 'subscribed';
       case AnalyticsTypes.unsubscribed:
-        return 'unsubscribed';
       case AnalyticsTypes.changed:
-        return 'changed';
       case AnalyticsTypes.denied:
-        return 'denied';
       case AnalyticsTypes.skipped:
-        return 'skipped';
       case AnalyticsTypes.checked:
-        return 'checked';
       case AnalyticsTypes.unchecked:
-        return 'unchecked';
       case AnalyticsTypes.attempted:
-        return 'attempted';
       case AnalyticsTypes.reset:
-        return 'reset';
       case AnalyticsTypes.enabled:
-        return 'enabled';
       case AnalyticsTypes.disabled:
-        return 'disabled';
       case AnalyticsTypes.began:
-        return 'began';
       case AnalyticsTypes.ended:
-        return 'ended';
       case AnalyticsTypes.refreshed:
-        return 'refreshed';
       case AnalyticsTypes.generated:
-        return 'generated';
       case AnalyticsTypes.unsupported:
-        return 'unsupported';
       case AnalyticsTypes.invalid:
-        return 'invalid';
       case AnalyticsTypes.valid:
-        return 'valid';
       case AnalyticsTypes.shown:
-        return 'shown';
       case AnalyticsTypes.loaded:
-        return 'loaded';
       case AnalyticsTypes.saved:
-        return 'saved';
+      case AnalyticsTypes.found:
+      case AnalyticsTypes.completed:
+      case AnalyticsTypes.error:
+      case AnalyticsTypes.given:
+        return name;
       case AnalyticsTypes.notFound:
         return 'not_found';
-      case AnalyticsTypes.found:
-        return 'found';
-      case AnalyticsTypes.completed:
-        return 'completed';
-      case AnalyticsTypes.error:
-        return 'error';
       case AnalyticsTypes.none:
         return '';
     }